refactor: update gitlab folder structure, and clean up
