+++ cmd/create.go
@@ -1,18 +1,13 @@
-/*
-Copyright © 2022 NAME HERE <EMAIL ADDRESS>
-
-*/
 package cmd
 
 import (
 	"fmt"
-	"log"
-	"time"
+	"github.com/kubefirst/nebulous/pkg/flare"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
-	"github.com/kubefirst/nebulous/pkg/flare"
+	"log"
+	"time"
 )
-
 
 const trackerStage20 = "0 - Apply Base"
 const trackerStage21 = "1 - Temporary SCM Install"
@@ -33,10 +28,10 @@ to quickly create a Cobra application.`,
 
 		flare.SetupProgress(4)
 		Trackers = make(map[string]*flare.ActionTracker)
-		Trackers[trackerStage20] = &flare.ActionTracker{flare.CreateTracker(trackerStage20, int64(1))}
-		Trackers[trackerStage21] = &flare.ActionTracker{flare.CreateTracker(trackerStage21, int64(2))}
-		Trackers[trackerStage22] = &flare.ActionTracker{flare.CreateTracker(trackerStage22, int64(7))}
-		Trackers[trackerStage23] = &flare.ActionTracker{flare.CreateTracker(trackerStage23, int64(3))}
+		Trackers[trackerStage20] = &flare.ActionTracker{Tracker: flare.CreateTracker(trackerStage20, int64(1))}
+		Trackers[trackerStage21] = &flare.ActionTracker{Tracker: flare.CreateTracker(trackerStage21, int64(2))}
+		Trackers[trackerStage22] = &flare.ActionTracker{Tracker: flare.CreateTracker(trackerStage22, int64(7))}
+		Trackers[trackerStage23] = &flare.ActionTracker{Tracker: flare.CreateTracker(trackerStage23, int64(3))}
 
 		infoCmd.Run(cmd, args)
 
@@ -48,46 +43,41 @@ to quickly create a Cobra application.`,
 		} else {
 			log.Printf("[#99] Dry-run mode, telemetry skipped:  %s", metricName)
 		}
-		
 
 		directory := fmt.Sprintf("%s/.kubefirst/gitops/terraform/base", home)
-		applyBaseTerraform(cmd,directory)
-		Trackers[trackerStage20].Tracker.Increment(int64(1))		
+		applyBaseTerraform(cmd, directory)
+		Trackers[trackerStage20].Tracker.Increment(int64(1))
 		createSoftServe(kubeconfigPath)
-		Trackers[trackerStage21].Tracker.Increment(int64(1))	
-		configureSoftserveAndPush()		
-		Trackers[trackerStage21].Tracker.Increment(int64(1))	
+		Trackers[trackerStage21].Tracker.Increment(int64(1))
+		configureSoftserveAndPush()
+		Trackers[trackerStage21].Tracker.Increment(int64(1))
 		helmInstallArgocd(home, kubeconfigPath)
 		Trackers[trackerStage22].Tracker.Increment(int64(1))
 
 		produceGitlabTokens()
 		Trackers[trackerStage22].Tracker.Increment(int64(1))
-		
-		
+
 		applyGitlabTerraform(directory)
 		Trackers[trackerStage22].Tracker.Increment(int64(1))
 		gitlabKeyUpload()
 		Trackers[trackerStage22].Tracker.Increment(int64(1))
-		
+
 		configureVault()
 		Trackers[trackerStage23].Tracker.Increment(int64(1))
-
 
 		addGitlabOidcApplications()
 		Trackers[trackerStage23].Tracker.Increment(int64(1))
 		awaitGitlab()
 		Trackers[trackerStage22].Tracker.Increment(int64(1))
-		
+
 		pushGitopsToGitLab()
 		Trackers[trackerStage22].Tracker.Increment(int64(1))
 		changeRegistryToGitLab()
 		Trackers[trackerStage22].Tracker.Increment(int64(1))
-		
-		
-		
+
 		hydrateGitlabMetaphorRepo()
 		Trackers[trackerStage23].Tracker.Increment(int64(1))
-		
+
 		token := getArgocdAuthToken()
 		syncArgocdApplication("argo-components", token)
 		syncArgocdApplication("gitlab-runner-components", token)
@@ -95,9 +85,8 @@ to quickly create a Cobra application.`,
 		syncArgocdApplication("atlantis-components", token)
 		syncArgocdApplication("chartmuseum-components", token)
 
-
 		metricName = "kubefirst.mgmt_cluster_install.completed"
-		
+
 		if !dryrunMode {
 			flare.SendTelemetry(metricDomain, metricName)
 		} else {
@@ -107,20 +96,10 @@ to quickly create a Cobra application.`,
 	},
 }
 
-
-
-
-
-
-
-
-
 func init() {
 	rootCmd.AddCommand(createCmd)
 
-	// createCmd.Flags().String("tf-entrypoint", "", "the entrypoint to execute the terraform from")
-	// createCmd.MarkFlagRequired("tf-entrypoint")
-	// todo make this an optional switch and check for it or viper
+	// todo: make this an optional switch and check for it or viper
 	createCmd.Flags().Bool("destroy", false, "destroy resources")
 	createCmd.PersistentFlags().BoolVarP(&dryrunMode, "dry-run", "s", false, "set to dry-run mode, no changes done on cloud provider selected")
